fix(): mainActivity增加跳转逻辑
