+++ app/src/main/java/com/example/ryan_ffmpeg_player/MainActivity.kt
@@ -1,5 +1,6 @@
 package com.example.ryan_ffmpeg_player
 
+import android.content.Intent
 import android.net.Uri
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
@@ -15,17 +16,11 @@ import java.io.File
 import java.lang.Exception
 
 class MainActivity : AppCompatActivity(), View.OnClickListener {
-    lateinit var playButton: Button
-    lateinit var videoSurfaceView: VideoSurfaceView
-    lateinit var albumButton: Button
-    lateinit var filePath: String
+    lateinit var jumpVideoButton: Button
 
-    private val pickLauncher =
-        registerForActivityResult(ActivityResultContracts.GetContent()) { uri: Uri? ->
-            uri?.let { it ->
-                filePath = FileUtils.getFilePathByUri(this, it)
-                Log.d(TAG, "图片地址：" + filePath)
-            }
+    private val launcher =
+        registerForActivityResult(ActivityResultContracts.StartActivityForResult()){ activityResult -> {
+        }
         }
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -33,68 +28,16 @@ class MainActivity : AppCompatActivity(), View.OnClickListener {
         setContentView(R.layout.activity_main)
 
         PermissionUtils.requestMyPermissions(this)
-        playButton = findViewById(R.id.playButton)
-        videoSurfaceView = findViewById(R.id.videoView)
-        albumButton = findViewById(R.id.sys_choose_image)
-        playButton.setOnClickListener(this)
-        albumButton.setOnClickListener(this)
-    }
-
-    override fun onResume() {
-        super.onResume()
-//        decode();
+        jumpVideoButton = findViewById(R.id.jump_video)
+        jumpVideoButton.setOnClickListener(this)
     }
 
     override fun onClick(v: View?) {
         when (v?.id) {
-            R.id.playButton -> {
-                if (!filePath.isEmpty()) {
-                    Log.d(TAG, "输入文件路径是否正常：" + File(filePath).exists())
-                    videoSurfaceView.player(filePath)
-                } else {
-                    Toast.makeText(this, "文件路径异常：" + filePath, Toast.LENGTH_LONG).show()
-                }
-            }
-            R.id.sys_choose_image -> {
-                // 图片选image/* 视频选video/*
-                pickLauncher.launch("video/*")
+            R.id.jump_video -> {
+                val intent = Intent().setAction("video_player")
+                launcher.launch(intent)
             }
         }
-    }
-
-    private fun decode() {
-        val path = "/sdcard/DCIM/Camera";
-        val fileName = "DCIM/ScreenRecorder/Screenrecorder-2022-11-09-10-52-55-338.mp4";
-        val input = File(
-            Environment.getExternalStorageDirectory(),
-            fileName
-        ).absolutePath
-        Log.e(TAG, "path" + Environment.getExternalStorageDirectory())
-
-        val output = File(
-            Environment.getExternalStorageDirectory(),
-            "test_yuv420p.yuv"
-        ).absolutePath
-
-        val outputFile = File(output)
-        if (outputFile.exists()) {
-            outputFile.delete();
-        }
-        try {
-            //创建文件
-            val createResult = outputFile.createNewFile();
-            //给一个吐司提示，提示创建成功
-        } catch (e: Exception) {
-            e.printStackTrace();
-            Log.e(TAG, "output e" + e)
-
-        }
-        Log.e(
-            TAG,
-            "input = $input ${File(input).exists()}, output = $output ${File(output).exists()}"
-        )
-
-        VideoUtil.decode(input, output)
-        VideoUtil.helloNDK()
     }
 }
