+++ app/src/main/res/layout/activity_main.xml
@@ -5,23 +5,12 @@
     android:orientation="vertical">
 
     <Button
-        android:id="@+id/sys_choose_image"
+        android:id="@+id/jump_video"
         android:layout_width="match_parent"
         android:layout_height="60dp"
         android:layout_marginVertical="16dp"
         android:layout_marginHorizontal="16dp"
         android:background="@color/black"
-        android:text="跳转到图片编辑展示页面"
+        android:text="跳转至视频播放页面"
         android:textColor="@color/white"/>
-
-    <Button
-        android:id="@+id/playButton"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="点击播放视频"/>
-
-    <com.example.ryan_ffmpeg_player.VideoSurfaceView
-        android:id="@+id/videoView"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
 </LinearLayout>
