+++ app/src/main/res/layout/activity_video_player.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    tools:context=".VideoPlayerActivity">
+
+    <Button
+        android:id="@+id/sys_choose_image"
+        android:layout_width="match_parent"
+        android:layout_height="60dp"
+        android:layout_marginVertical="16dp"
+        android:layout_marginHorizontal="16dp"
+        android:background="@color/black"
+        android:text="跳转到图片编辑展示页面"
+        android:textColor="@color/white"/>
+
+    <Button
+        android:id="@+id/playButton"
+        android:layout_width="match_parent"
+        android:layout_height="60dp"
+        android:layout_marginVertical="16dp"
+        android:layout_marginHorizontal="16dp"
+        android:background="@color/black"
+        android:text="点击播放视频"
+        android:textColor="@color/white"/>
+
+    <com.example.ryan_ffmpeg_player.VideoSurfaceView
+        android:id="@+id/videoView"
+        android:layout_width="0dp"
+        android:layout_height="0dp"/>
+</LinearLayout>
