+++ app/src/main/res/xml/fraction.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <item name="circle_article_aspectRatio" type="fraction">133%</item>
+</resources>
